Sun, May 26, 2024, 9:22 PM +05:30

+++ readability/readability.c
@@ -81,6 +81,6 @@ int sentance_count(string text)
 
 int grade_count(int L, int S)
 {
-    float grade = roundf(0.0588 * L - 0.296 * S - 15.8);
+    float grade = round(0.0588 * L - 0.296 * S - 15.8);
     return grade;
 }
